@@ -6,14 +6,13 @@ publish_to: 'none'
 version: 1.1.4+104
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_svg: ^1.0.3
   flutter_secure_storage: ^5.0.2
   firebase_messaging: any
   firebase_core: any
@@ -21,7 +20,7 @@ dependencies:
   http: ^0.13.4
   auto_size_text: 3.0.0
   intl: ^0.17.0
-  fl_chart: ^0.50.1
+  fl_chart: ^0.50.5
   animations: ^2.0.2
   progress_indicators: ^1.0.0
   url_launcher: ^6.1.2
@@ -62,7 +61,7 @@ dependencies:
   cupertino_icons: ^1.0.4
 
 dev_dependencies:
-  flutter_lints: ^1.0.4
+  flutter_lints: ^2.0.1
   flutter_launcher_name: ^0.0.1
   flutter_test:
     sdk: flutter
