@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
@@ -28,10 +29,10 @@ class LoginScreen extends StatefulWidget {
   const LoginScreen({Key? key}) : super(key: key);
 
   @override
-  _LoginScreenState createState() => _LoginScreenState();
+  LoginScreenState createState() => LoginScreenState();
 }
 
-class _LoginScreenState extends State<LoginScreen> {
+class LoginScreenState extends State<LoginScreen> {
   PackageInfo? _packageInfo;
   TextEditingController loginController = TextEditingController();
   TextEditingController passwordController = TextEditingController();
@@ -57,7 +58,9 @@ class _LoginScreenState extends State<LoginScreen> {
   void _handleStuff() async {
     String? s = await SecureStorage.readStorage(key: "next");
     if (s == null) {
-      print("NULL");
+      if (kDebugMode) {
+        print("NULL");
+      }
       await const FlutterSecureStorage().deleteAll();
       await SecureStorage.writeStorage(key: "next", value: "1");
     }
@@ -106,26 +109,25 @@ class _LoginScreenState extends State<LoginScreen> {
       setState(() {
         _curtain = false;
       });
-      ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-        content: Text(
-          "Invalid credentials",
-          textAlign: TextAlign.center,
-        ),
-        backgroundColor: Colors.red,
-        behavior: SnackBarBehavior.fixed,
-        elevation: 5.0,
-      ));
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+          content: Text(
+            "Invalid credentials",
+            textAlign: TextAlign.center,
+          ),
+          backgroundColor: Colors.red,
+          behavior: SnackBarBehavior.fixed,
+          elevation: 5.0,
+        ));
+      }
     }
   }
 
   void _loginUser(String login, String pass) async {
     var email = login;
-    bool emailValid = RegExp(
-            r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
-        .hasMatch(email);
+    bool emailValid = RegExp(r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+").hasMatch(email);
     if (!emailValid) {
-      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.email_invalid,
-          AppLocalizations.of(context)!.email_invalid_message);
+      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.email_invalid, AppLocalizations.of(context)!.email_invalid_message);
       setState(() {
         _curtain = false;
       });
@@ -138,32 +140,36 @@ class _LoginScreenState extends State<LoginScreen> {
     if (res != null) {
       bool code = await NetInterface.getEmailCode(res);
       if (code) {
-        Dialogs.open2FAbox(context, res, _getToken);
+        if (mounted) Dialogs.open2FAbox(context, res, _getToken);
       } else {
-        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-          content: Text(
-            "Problem getting your email code, try again later",
-            textAlign: TextAlign.center,
-          ),
-          backgroundColor: Colors.red,
-          behavior: SnackBarBehavior.fixed,
-          elevation: 5.0,
-        ));
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+            content: Text(
+              "Problem getting your email code, try again later",
+              textAlign: TextAlign.center,
+            ),
+            backgroundColor: Colors.red,
+            behavior: SnackBarBehavior.fixed,
+            elevation: 5.0,
+          ));
+        }
         setState(() {
           _curtain = false;
         });
       }
       // _codeDialog(res);
     } else {
-      ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-        content: Text(
-          "Credentials doesn't match any user!",
-          textAlign: TextAlign.center,
-        ),
-        backgroundColor: Colors.red,
-        behavior: SnackBarBehavior.fixed,
-        elevation: 5.0,
-      ));
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+          content: Text(
+            "Credentials doesn't match any user!",
+            textAlign: TextAlign.center,
+          ),
+          backgroundColor: Colors.red,
+          behavior: SnackBarBehavior.fixed,
+          elevation: 5.0,
+        ));
+      }
       setState(() {
         _curtain = false;
       });
@@ -176,8 +182,9 @@ class _LoginScreenState extends State<LoginScreen> {
       await SecureStorage.writeStorage(key: NetInterface.token, value: res);
       String? wer = await SecureStorage.readStorage(key: NetInterface.token);
       if (wer != null) {
-        Navigator.pushReplacement(context,
-            MaterialPageRoute(builder: (context) => const PortfolioScreen()));
+        if (mounted) {
+          Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => const PortfolioScreen()));
+        }
       }
     } else {
       if (mounted) {
@@ -201,21 +208,21 @@ class _LoginScreenState extends State<LoginScreen> {
     String? res = await SecureStorage.readStorage(key: "PIN");
     // String? res;
     if (res == null) {
-      Navigator.of(context).pushReplacement(PageRouteBuilder(
-          pageBuilder: (BuildContext context, _, __) {
+      if (mounted) {
+        Navigator.of(context).pushReplacement(PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
         return const PortfolioScreen();
-      }, transitionsBuilder:
-              (_, Animation<double> animation, __, Widget child) {
+      }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
         return FadeTransition(opacity: animation, child: child);
       }));
+      }
     } else {
-      Navigator.of(context).pushReplacement(PageRouteBuilder(
-          pageBuilder: (BuildContext context, _, __) {
+      if (mounted) {
+        Navigator.of(context).pushReplacement(PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
         return const AuthScreen();
-      }, transitionsBuilder:
-              (_, Animation<double> animation, __, Widget child) {
+      }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
         return FadeTransition(opacity: animation, child: child);
       }));
+      }
     }
   }
 
@@ -243,39 +250,25 @@ class _LoginScreenState extends State<LoginScreen> {
     var password = passwordRegController.text;
     var passwordConfirm = passwordRegConfirmController.text;
     var email = emailRegController.text;
-    bool emailValid = RegExp(
-            r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
-        .hasMatch(email);
+    bool emailValid = RegExp(r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+").hasMatch(email);
 
     if (username.isEmpty || username.length > 64) {
-      Dialogs.openAlertBox(
-          context,
-          AppLocalizations.of(context)!.username_invalid,
-          AppLocalizations.of(context)!.username_invalid_message);
+      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.username_invalid, AppLocalizations.of(context)!.username_invalid_message);
       return;
     } else if (password.length < 8 || password.length > 64) {
-      Dialogs.openAlertBox(
-          context,
-          AppLocalizations.of(context)!.password_invalid,
-          AppLocalizations.of(context)!.password_invalid_message);
+      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.password_invalid, AppLocalizations.of(context)!.password_invalid_message);
       return;
     } else if (!emailValid) {
-      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.email_invalid,
-          AppLocalizations.of(context)!.email_invalid_message);
+      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.email_invalid, AppLocalizations.of(context)!.email_invalid_message);
       return;
     } else if (realname.isEmpty || realname.length > 64) {
-      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.name_invalid,
-          AppLocalizations.of(context)!.name_invalid_message);
+      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.name_invalid, AppLocalizations.of(context)!.name_invalid_message);
       return;
     } else if (password != passwordConfirm) {
-      Dialogs.openAlertBox(
-          context,
-          AppLocalizations.of(context)!.password_mismatch,
-          AppLocalizations.of(context)!.password_mismatch_message);
+      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.password_mismatch, AppLocalizations.of(context)!.password_mismatch_message);
       return;
     } else if (!_termsAgreed) {
-      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.alert,
-          AppLocalizations.of(context)!.terms_agree);
+      Dialogs.openAlertBox(context, AppLocalizations.of(context)!.alert, AppLocalizations.of(context)!.terms_agree);
       return;
     }
     setState(() {
@@ -304,7 +297,7 @@ class _LoginScreenState extends State<LoginScreen> {
     } else {
       var m = RegistrationErrors.fromJson(error);
       // print(m.status);
-      Dialogs.openAlertBox(context, m.title!, m.errors.toString());
+      if (mounted) Dialogs.openAlertBox(context, m.title!, m.errors.toString());
       setState(() {
         _curtain = false;
         _registerButton = true;
@@ -313,7 +306,7 @@ class _LoginScreenState extends State<LoginScreen> {
   }
 
   void _forgotPass(String email) async {
-    var i = NetInterface.forgotPass(email);
+    NetInterface.forgotPass(email);
   }
 
   @override
@@ -327,11 +320,8 @@ class _LoginScreenState extends State<LoginScreen> {
                 child: Padding(
               padding: const EdgeInsets.only(right: 15.0, top: 17.0),
               child: Text(
-                'v ' + _appVersion,
-                style: Theme.of(context)
-                    .textTheme
-                    .headline4!
-                    .copyWith(color: Colors.white70, fontSize: 12.0),
+                'v $_appVersion',
+                style: Theme.of(context).textTheme.headline4!.copyWith(color: Colors.white70, fontSize: 12.0),
               ),
             )),
           ),
@@ -344,281 +334,204 @@ class _LoginScreenState extends State<LoginScreen> {
                   alignment: Alignment.topCenter,
                   child: SafeArea(
                     child: SingleChildScrollView(
-                      child: Column(
-                          mainAxisAlignment: MainAxisAlignment.start,
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            const SizedBox(
-                              height: 150,
-                            ),
-                            Align(
-                              alignment: Alignment.center,
-                              child: SizedBox(
-                                width: 280,
-                                child: Column(
-                                  mainAxisSize: MainAxisSize.min,
-                                  mainAxisAlignment: MainAxisAlignment.end,
-                                  crossAxisAlignment: CrossAxisAlignment.end,
-                                  children: [
-                                    SizedBox(
-                                      width: 280,
-                                      child: NeuContainer(
-                                          child: TextField(
-                                              style: Theme.of(context)
-                                                  .textTheme
-                                                  .bodyText1!
-                                                  .copyWith(
-                                                      color: Colors.white,
-                                                      fontSize: 18.0),
-                                              autocorrect: false,
-                                              controller: loginController,
-                                              textAlign: TextAlign.center,
-                                              decoration: InputDecoration(
-                                                isDense: false,
-                                                contentPadding:
-                                                    const EdgeInsets.only(
-                                                        left: 10.0,
-                                                        bottom: 5.0),
-                                                hintStyle: Theme.of(context)
-                                                    .textTheme
-                                                    .subtitle1!
-                                                    .copyWith(
-                                                        color: Colors.white54,
-                                                        fontSize: 14.0),
-                                                hintText: AppLocalizations.of(
-                                                        context)!
-                                                    .e_mail,
-                                                enabledBorder:
-                                                    const UnderlineInputBorder(
-                                                  borderSide: BorderSide(
-                                                      color:
-                                                          Colors.transparent),
-                                                ),
-                                                focusedBorder:
-                                                    const UnderlineInputBorder(
-                                                  borderSide: BorderSide(
-                                                      color:
-                                                          Colors.transparent),
-                                                ),
-                                              ))),
-                                    ),
-                                    const SizedBox(
-                                      height: 30.0,
-                                    ),
-                                    SizedBox(
-                                      width: 280,
-                                      child: NeuContainer(
-                                          child: TextField(
-                                              style: Theme.of(context)
-                                                  .textTheme
-                                                  .bodyText1!
-                                                  .copyWith(
-                                                      color: Colors.white,
-                                                      fontSize: 18.0),
-                                              obscureText: true,
-                                              enableSuggestions: false,
-                                              autocorrect: false,
-                                              controller: passwordController,
-                                              textAlign: TextAlign.center,
-                                              decoration: InputDecoration(
-                                                isDense: false,
-                                                contentPadding:
-                                                    const EdgeInsets.only(
-                                                        left: 10.0,
-                                                        bottom: 5.0),
-                                                hintStyle: Theme.of(context)
-                                                    .textTheme
-                                                    .subtitle1!
-                                                    .copyWith(
-                                                        color: Colors.white54,
-                                                        fontSize: 14.0),
-                                                hintText: AppLocalizations.of(
-                                                        context)!
-                                                    .password,
-                                                enabledBorder:
-                                                    const UnderlineInputBorder(
-                                                  borderSide: BorderSide(
-                                                      color:
-                                                          Colors.transparent),
-                                                ),
-                                                focusedBorder:
-                                                    const UnderlineInputBorder(
-                                                  borderSide: BorderSide(
-                                                      color:
-                                                          Colors.transparent),
-                                                ),
-                                              ))),
-                                    ),
-                                    const SizedBox(
-                                      height: 30.0,
-                                    ),
-                                    Padding(
-                                      padding:
-                                          const EdgeInsets.only(left: 55.0),
-                                      child: Row(
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        children: [
-                                          Text(
-                                            AppLocalizations.of(context)!
-                                                .forgot_pass,
-                                            style: Theme.of(context)
-                                                .textTheme
-                                                .subtitle1,
-                                          ),
-                                          const SizedBox(
-                                            width: 20.0,
-                                          ),
-                                          SizedBox(
-                                            height: 30,
-                                            width: 25,
-                                            child: NeuButton(
-                                              onTap: () {
-                                                _forgotDialog();
-                                              },
-                                              icon: const Icon(
-                                                Icons.arrow_forward_ios,
-                                                size: 20.0,
-                                                color: Colors.white70,
-                                              ),
+                      child: Column(mainAxisAlignment: MainAxisAlignment.start, crossAxisAlignment: CrossAxisAlignment.center, children: [
+                        const SizedBox(
+                          height: 150,
+                        ),
+                        Align(
+                          alignment: Alignment.center,
+                          child: SizedBox(
+                            width: 280,
+                            child: Column(
+                              mainAxisSize: MainAxisSize.min,
+                              mainAxisAlignment: MainAxisAlignment.end,
+                              crossAxisAlignment: CrossAxisAlignment.end,
+                              children: [
+                                SizedBox(
+                                  width: 280,
+                                  child: NeuContainer(
+                                      child: TextField(
+                                          style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
+                                          autocorrect: false,
+                                          controller: loginController,
+                                          textAlign: TextAlign.center,
+                                          decoration: InputDecoration(
+                                            isDense: false,
+                                            contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                            hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                            hintText: AppLocalizations.of(context)!.e_mail,
+                                            enabledBorder: const UnderlineInputBorder(
+                                              borderSide: BorderSide(color: Colors.transparent),
                                             ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                  ],
+                                            focusedBorder: const UnderlineInputBorder(
+                                              borderSide: BorderSide(color: Colors.transparent),
+                                            ),
+                                          ))),
                                 ),
-                              ),
-                            ),
-                            const SizedBox(
-                              height: 50,
-                            ),
-                            SizedBox(
-                              width: 250,
-                              child: NeuButton(
-                                onTap: () {
-                                  _loginUser(loginController.text,
-                                      passwordController.text);
-                                },
-                                splashColor: Colors.purple,
-                                child: Container(
-                                  width: 200,
-                                  height: 50,
-                                  color: Colors.transparent,
-                                  child: Center(
-                                      child: Text(
-                                    AppLocalizations.of(context)!.sign_in,
-                                    style: Theme.of(context)
-                                        .textTheme
-                                        .bodyText1!
-                                        .copyWith(
-                                            fontSize: 22.0,
-                                            color: Colors.white),
-                                  )),
+                                const SizedBox(
+                                  height: 30.0,
                                 ),
-                              ),
-                            ),
-                            const SizedBox(
-                              height: 20.0,
-                            ),
-                            Text(
-                              AppLocalizations.of(context)!.or,
-                              style: Theme.of(context)
-                                  .textTheme
-                                  .subtitle1!
-                                  .copyWith(
-                                      fontSize: 16.0, color: Colors.white),
-                            ),
-                            const SizedBox(
-                              height: 20.0,
-                            ),
-                            SizedBox(
-                              width: 250,
-                              height: 50,
-                              child: NeuButton(
-                                  onTap: () async {
-                                    FirebaseService service = FirebaseService();
-                                    try {
-                                      String? tokenID =
-                                          await service.signInwithGoogle();
-                                      setState(() {
-                                        _curtain = true;
-                                      });
-                                      if (tokenID != null) {
-                                        var asdf =
-                                            await NetInterface.getTokenGoogle(
-                                                tokenID);
-                                        if (asdf != null) {
-                                          _nextPage();
-                                        } else {
-                                          Dialogs.openAlertBox(context, "Error",
-                                              "Error Sign in with Google");
-                                          setState(() {
-                                            _curtain = false;
-                                          });
-                                        }
-                                      }
-                                    } catch (e) {
-                                      print("======HOVNO=======");
-                                      print(e);
-                                      if (e is FirebaseAuthException) {
-                                        print(e.message!);
-                                      }
-                                      Dialogs.openAlertBox(context, "Error",
-                                          "Error Sign in with Google");
-                                    }
-                                    setState(() {
-                                      // isLoading = false;
-                                    });
-                                  },
-                                  splashColor: Colors.purple,
-                                  child: Wrap(
-                                    crossAxisAlignment:
-                                        WrapCrossAlignment.center,
+                                SizedBox(
+                                  width: 280,
+                                  child: NeuContainer(
+                                      child: TextField(
+                                          style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
+                                          obscureText: true,
+                                          enableSuggestions: false,
+                                          autocorrect: false,
+                                          controller: passwordController,
+                                          textAlign: TextAlign.center,
+                                          decoration: InputDecoration(
+                                            isDense: false,
+                                            contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                            hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                            hintText: AppLocalizations.of(context)!.password,
+                                            enabledBorder: const UnderlineInputBorder(
+                                              borderSide: BorderSide(color: Colors.transparent),
+                                            ),
+                                            focusedBorder: const UnderlineInputBorder(
+                                              borderSide: BorderSide(color: Colors.transparent),
+                                            ),
+                                          ))),
+                                ),
+                                const SizedBox(
+                                  height: 30.0,
+                                ),
+                                Padding(
+                                  padding: const EdgeInsets.only(left: 55.0),
+                                  child: Row(
+                                    mainAxisAlignment: MainAxisAlignment.center,
                                     children: [
-                                      SizedBox(
-                                          height: 25,
-                                          child: Image.asset(
-                                              'images/google_icon.png')),
-                                      const SizedBox(
-                                        width: 10.0,
-                                      ),
                                       Text(
-                                        AppLocalizations.of(context)!
-                                            .sign_in_google,
-                                        style: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1!
-                                            .copyWith(
-                                                fontSize: 14.0,
-                                                color: Colors.white),
+                                        AppLocalizations.of(context)!.forgot_pass,
+                                        style: Theme.of(context).textTheme.subtitle1,
+                                      ),
+                                      const SizedBox(
+                                        width: 20.0,
+                                      ),
+                                      SizedBox(
+                                        height: 30,
+                                        width: 25,
+                                        child: NeuButton(
+                                          onTap: () {
+                                            _forgotDialog();
+                                          },
+                                          icon: const Icon(
+                                            Icons.arrow_forward_ios,
+                                            size: 20.0,
+                                            color: Colors.white70,
+                                          ),
+                                        ),
                                       ),
                                     ],
-                                  )),
+                                  ),
+                                ),
+                              ],
                             ),
-                            const SizedBox(
-                              height: 20.0,
+                          ),
+                        ),
+                        const SizedBox(
+                          height: 50,
+                        ),
+                        SizedBox(
+                          width: 250,
+                          child: NeuButton(
+                            onTap: () {
+                              _loginUser(loginController.text, passwordController.text);
+                            },
+                            splashColor: Colors.purple,
+                            child: Container(
+                              width: 200,
+                              height: 50,
+                              color: Colors.transparent,
+                              child: Center(
+                                  child: Text(
+                                AppLocalizations.of(context)!.sign_in,
+                                style: Theme.of(context).textTheme.bodyText1!.copyWith(fontSize: 22.0, color: Colors.white),
+                              )),
                             ),
-                            Platform.isIOS
-                                ? AppleSignInButton(
-                                    onSignIn: (AuthorizationCredentialAppleID
-                                        value) async {
-                                      var asdf =
-                                          await NetInterface.getTokenApple(
-                                              value.authorizationCode);
-                                      if (asdf != null) {
-                                        _nextPage();
-                                      } else {
-                                        Dialogs.openAlertBox(context, "Error",
-                                            "Error Sign in with Apple");
-                                        setState(() {
-                                          _curtain = false;
-                                        });
-                                      }
-                                    },
-                                  )
-                                : Container(),
-                          ]),
+                          ),
+                        ),
+                        const SizedBox(
+                          height: 20.0,
+                        ),
+                        Text(
+                          AppLocalizations.of(context)!.or,
+                          style: Theme.of(context).textTheme.subtitle1!.copyWith(fontSize: 16.0, color: Colors.white),
+                        ),
+                        const SizedBox(
+                          height: 20.0,
+                        ),
+                        SizedBox(
+                          width: 250,
+                          height: 50,
+                          child: NeuButton(
+                              onTap: () async {
+                                FirebaseService service = FirebaseService();
+                                try {
+                                  String? tokenID = await service.signInwithGoogle();
+                                  setState(() {
+                                    _curtain = true;
+                                  });
+                                  if (tokenID != null) {
+                                    var asdf = await NetInterface.getTokenGoogle(tokenID);
+                                    if (asdf != null) {
+                                      _nextPage();
+                                    } else {
+                                      if (mounted) Dialogs.openAlertBox(context, "Error", "Error Sign in with Google");
+                                      setState(() {
+                                        _curtain = false;
+                                      });
+                                    }
+                                  }
+                                } catch (e) {
+                                  if (kDebugMode) {
+                                    print("======HOVNO=======");
+                                    print(e);
+                                  }
+                                  if (e is FirebaseAuthException) {
+                                    if (kDebugMode) print(e.message!);
+                                  }
+                                  Dialogs.openAlertBox(context, "Error", "Error Sign in with Google");
+                                }
+                                setState(() {
+                                  // isLoading = false;
+                                });
+                              },
+                              splashColor: Colors.purple,
+                              child: Wrap(
+                                crossAxisAlignment: WrapCrossAlignment.center,
+                                children: [
+                                  SizedBox(height: 25, child: Image.asset('images/google_icon.png')),
+                                  const SizedBox(
+                                    width: 10.0,
+                                  ),
+                                  Text(
+                                    AppLocalizations.of(context)!.sign_in_google,
+                                    style: Theme.of(context).textTheme.subtitle1!.copyWith(fontSize: 14.0, color: Colors.white),
+                                  ),
+                                ],
+                              )),
+                        ),
+                        const SizedBox(
+                          height: 20.0,
+                        ),
+                        Platform.isIOS
+                            ? AppleSignInButton(
+                                onSignIn: (AuthorizationCredentialAppleID value) async {
+                                  var asdf = await NetInterface.getTokenApple(value.authorizationCode);
+                                  if (asdf != null) {
+                                    _nextPage();
+                                  } else {
+                                    if (mounted) Dialogs.openAlertBox(context, "Error", "Error Sign in with Apple");
+                                    setState(() {
+                                      _curtain = false;
+                                    });
+                                  }
+                                },
+                              )
+                            : Container(),
+                      ]),
                     ),
                   )),
             ),
@@ -632,299 +545,196 @@ class _LoginScreenState extends State<LoginScreen> {
                   alignment: Alignment.topCenter,
                   child: SafeArea(
                     child: SingleChildScrollView(
-                      child: Column(
-                          mainAxisAlignment: MainAxisAlignment.start,
-                          crossAxisAlignment: CrossAxisAlignment.center,
+                      child: Column(mainAxisAlignment: MainAxisAlignment.start, crossAxisAlignment: CrossAxisAlignment.center, children: [
+                        const SizedBox(
+                          height: 150,
+                        ),
+                        SizedBox(
+                          width: 280,
+                          child: NeuContainer(
+                              child: TextField(
+                                  style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
+                                  autocorrect: false,
+                                  controller: emailRegController,
+                                  textAlign: TextAlign.center,
+                                  decoration: InputDecoration(
+                                    isDense: false,
+                                    contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                    hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                    hintText: AppLocalizations.of(context)!.e_mail,
+                                    enabledBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                    focusedBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                  ))),
+                        ),
+                        const SizedBox(
+                          height: 30.0,
+                        ),
+                        SizedBox(
+                          width: 280,
+                          child: NeuContainer(
+                              child: TextField(
+                                  style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
+                                  autocorrect: false,
+                                  controller: firstNameController,
+                                  textAlign: TextAlign.center,
+                                  decoration: InputDecoration(
+                                    isDense: false,
+                                    contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                    hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                    hintText: AppLocalizations.of(context)!.name,
+                                    enabledBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                    focusedBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                  ))),
+                        ),
+                        const SizedBox(
+                          height: 30.0,
+                        ),
+                        SizedBox(
+                          width: 280,
+                          child: NeuContainer(
+                              child: TextField(
+                                  style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
+                                  autocorrect: false,
+                                  controller: secondNameController,
+                                  textAlign: TextAlign.center,
+                                  decoration: InputDecoration(
+                                    isDense: false,
+                                    contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                    hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                    hintText: AppLocalizations.of(context)!.surname,
+                                    enabledBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                    focusedBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                  ))),
+                        ),
+                        const SizedBox(
+                          height: 30.0,
+                        ),
+                        SizedBox(
+                          width: 280,
+                          child: NeuContainer(
+                              child: TextField(
+                                  style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
+                                  obscureText: true,
+                                  enableSuggestions: false,
+                                  autocorrect: false,
+                                  controller: passwordRegController,
+                                  textAlign: TextAlign.center,
+                                  decoration: InputDecoration(
+                                    isDense: false,
+                                    contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                    hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                    hintText: AppLocalizations.of(context)!.password,
+                                    enabledBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                    focusedBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                  ))),
+                        ),
+                        const SizedBox(
+                          height: 30.0,
+                        ),
+                        SizedBox(
+                          width: 280,
+                          child: NeuContainer(
+                              child: TextField(
+                                  style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
+                                  obscureText: true,
+                                  enableSuggestions: false,
+                                  autocorrect: false,
+                                  controller: passwordRegConfirmController,
+                                  textAlign: TextAlign.center,
+                                  decoration: InputDecoration(
+                                    isDense: false,
+                                    contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                    hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                    hintText: AppLocalizations.of(context)!.conf_password,
+                                    enabledBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                    focusedBorder: const UnderlineInputBorder(
+                                      borderSide: BorderSide(color: Colors.transparent),
+                                    ),
+                                  ))),
+                        ),
+                        const SizedBox(
+                          height: 30.0,
+                        ),
+                        Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
                           children: [
-                            const SizedBox(
-                              height: 150,
-                            ),
-                            SizedBox(
-                              width: 280,
-                              child: NeuContainer(
-                                  child: TextField(
-                                      style: Theme.of(context)
-                                          .textTheme
-                                          .bodyText1!
-                                          .copyWith(
-                                              color: Colors.white,
-                                              fontSize: 18.0),
-                                      autocorrect: false,
-                                      controller: emailRegController,
-                                      textAlign: TextAlign.center,
-                                      decoration: InputDecoration(
-                                        isDense: false,
-                                        contentPadding: const EdgeInsets.only(
-                                            left: 10.0, bottom: 5.0),
-                                        hintStyle: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1!
-                                            .copyWith(
-                                                color: Colors.white54,
-                                                fontSize: 14.0),
-                                        hintText: AppLocalizations.of(context)!
-                                            .e_mail,
-                                        enabledBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                        focusedBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                      ))),
-                            ),
-                            const SizedBox(
-                              height: 30.0,
-                            ),
-                            SizedBox(
-                              width: 280,
-                              child: NeuContainer(
-                                  child: TextField(
-                                      style: Theme.of(context)
-                                          .textTheme
-                                          .bodyText1!
-                                          .copyWith(
-                                              color: Colors.white,
-                                              fontSize: 18.0),
-                                      autocorrect: false,
-                                      controller: firstNameController,
-                                      textAlign: TextAlign.center,
-                                      decoration: InputDecoration(
-                                        isDense: false,
-                                        contentPadding: const EdgeInsets.only(
-                                            left: 10.0, bottom: 5.0),
-                                        hintStyle: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1!
-                                            .copyWith(
-                                                color: Colors.white54,
-                                                fontSize: 14.0),
-                                        hintText:
-                                            AppLocalizations.of(context)!.name,
-                                        enabledBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                        focusedBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                      ))),
-                            ),
-                            const SizedBox(
-                              height: 30.0,
-                            ),
-                            SizedBox(
-                              width: 280,
-                              child: NeuContainer(
-                                  child: TextField(
-                                      style: Theme.of(context)
-                                          .textTheme
-                                          .bodyText1!
-                                          .copyWith(
-                                              color: Colors.white,
-                                              fontSize: 18.0),
-                                      autocorrect: false,
-                                      controller: secondNameController,
-                                      textAlign: TextAlign.center,
-                                      decoration: InputDecoration(
-                                        isDense: false,
-                                        contentPadding: const EdgeInsets.only(
-                                            left: 10.0, bottom: 5.0),
-                                        hintStyle: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1!
-                                            .copyWith(
-                                                color: Colors.white54,
-                                                fontSize: 14.0),
-                                        hintText: AppLocalizations.of(context)!
-                                            .surname,
-                                        enabledBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                        focusedBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                      ))),
-                            ),
-                            const SizedBox(
-                              height: 30.0,
-                            ),
-                            SizedBox(
-                              width: 280,
-                              child: NeuContainer(
-                                  child: TextField(
-                                      style: Theme.of(context)
-                                          .textTheme
-                                          .bodyText1!
-                                          .copyWith(
-                                              color: Colors.white,
-                                              fontSize: 18.0),
-                                      obscureText: true,
-                                      enableSuggestions: false,
-                                      autocorrect: false,
-                                      controller: passwordRegController,
-                                      textAlign: TextAlign.center,
-                                      decoration: InputDecoration(
-                                        isDense: false,
-                                        contentPadding: const EdgeInsets.only(
-                                            left: 10.0, bottom: 5.0),
-                                        hintStyle: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1!
-                                            .copyWith(
-                                                color: Colors.white54,
-                                                fontSize: 14.0),
-                                        hintText: AppLocalizations.of(context)!
-                                            .password,
-                                        enabledBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                        focusedBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                      ))),
-                            ),
-                            const SizedBox(
-                              height: 30.0,
-                            ),
-                            SizedBox(
-                              width: 280,
-                              child: NeuContainer(
-                                  child: TextField(
-                                      style: Theme.of(context)
-                                          .textTheme
-                                          .bodyText1!
-                                          .copyWith(
-                                              color: Colors.white,
-                                              fontSize: 18.0),
-                                      obscureText: true,
-                                      enableSuggestions: false,
-                                      autocorrect: false,
-                                      controller: passwordRegConfirmController,
-                                      textAlign: TextAlign.center,
-                                      decoration: InputDecoration(
-                                        isDense: false,
-                                        contentPadding: const EdgeInsets.only(
-                                            left: 10.0, bottom: 5.0),
-                                        hintStyle: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1!
-                                            .copyWith(
-                                                color: Colors.white54,
-                                                fontSize: 14.0),
-                                        hintText: AppLocalizations.of(context)!
-                                            .conf_password,
-                                        enabledBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                        focusedBorder:
-                                            const UnderlineInputBorder(
-                                          borderSide: BorderSide(
-                                              color: Colors.transparent),
-                                        ),
-                                      ))),
-                            ),
-                            const SizedBox(
-                              height: 30.0,
-                            ),
-                            Column(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: [
-                                RichText(
-                                  text: TextSpan(
-                                    children: [
-                                      TextSpan(
-                                        text: AppLocalizations.of(context)!
-                                                .agreed_to +
-                                            ' ',
-                                        style: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1,
-                                      ),
-                                      TextSpan(
-                                        text:
-                                            AppLocalizations.of(context)!.terms,
-                                        style: Theme.of(context)
-                                            .textTheme
-                                            .subtitle1!
-                                            .copyWith(color: Colors.blue),
-                                        recognizer: TapGestureRecognizer()
-                                          ..onTap = () {
-                                            _launchURL(
-                                                'https://rocketbot.pro/terms');
-                                          },
-                                      ),
-                                    ],
+                            RichText(
+                              text: TextSpan(
+                                children: [
+                                  TextSpan(
+                                    text: '${AppLocalizations.of(context)!.agreed_to} ',
+                                    style: Theme.of(context).textTheme.subtitle1,
                                   ),
-                                ),
-                                const SizedBox(
-                                  height: 5.0,
-                                ),
-                                Container(
-                                  height: 25,
-                                  width: 25,
-                                  decoration: BoxDecoration(
-                                    color: Colors.white.withOpacity(0.05),
-                                    borderRadius: const BorderRadius.all(
-                                        Radius.circular(5.0)),
+                                  TextSpan(
+                                    text: AppLocalizations.of(context)!.terms,
+                                    style: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.blue),
+                                    recognizer: TapGestureRecognizer()
+                                      ..onTap = () {
+                                        _launchURL('https://rocketbot.pro/terms');
+                                      },
                                   ),
-                                  child: Checkbox(
-                                      checkColor: Colors.lightGreen,
-                                      activeColor: Colors.white12,
-                                      value: _termsAgreed,
-                                      onChanged: _onTermsChanged),
-                                ),
-                              ],
+                                ],
+                              ),
                             ),
                             const SizedBox(
-                              height: 45,
+                              height: 5.0,
                             ),
-                            SizedBox(
-                                width: 250,
-                                height: 50,
-                                child: _registerButton
-                                    ? NeuButton(
-                                        onTap: () {
-                                          _registerUser();
-                                        },
-                                        splashColor: Colors.purple,
-                                        child: GradientText(
-                                          AppLocalizations.of(context)!
-                                              .register_button,
-                                          gradient:
-                                              const LinearGradient(colors: [
-                                            Color(0xFFF05523),
-                                            Color(0xFF812D88),
-                                          ]),
-                                          style: Theme.of(context)
-                                              .textTheme
-                                              .bodyText1!
-                                              .copyWith(
-                                                  fontSize: 22.0,
-                                                  color: Colors.white),
-                                        ))
-                                    : const Center(
-                                        child: CircularProgressIndicator(
-                                        strokeWidth: 2.0,
-                                        color: Color(0xFFAA3B63),
-                                      ))),
-                          ]),
+                            Container(
+                              height: 25,
+                              width: 25,
+                              decoration: BoxDecoration(
+                                color: Colors.white.withOpacity(0.05),
+                                borderRadius: const BorderRadius.all(Radius.circular(5.0)),
+                              ),
+                              child: Checkbox(
+                                  checkColor: Colors.lightGreen, activeColor: Colors.white12, value: _termsAgreed, onChanged: _onTermsChanged),
+                            ),
+                          ],
+                        ),
+                        const SizedBox(
+                          height: 45,
+                        ),
+                        SizedBox(
+                            width: 250,
+                            height: 50,
+                            child: _registerButton
+                                ? NeuButton(
+                                    onTap: () {
+                                      _registerUser();
+                                    },
+                                    splashColor: Colors.purple,
+                                    child: GradientText(
+                                      AppLocalizations.of(context)!.register_button,
+                                      gradient: const LinearGradient(colors: [
+                                        Color(0xFFF05523),
+                                        Color(0xFF812D88),
+                                      ]),
+                                      style: Theme.of(context).textTheme.bodyText1!.copyWith(fontSize: 22.0, color: Colors.white),
+                                    ))
+                                : const Center(
+                                    child: CircularProgressIndicator(
+                                    strokeWidth: 2.0,
+                                    color: Color(0xFFAA3B63),
+                                  ))),
+                      ]),
                     ),
                   )),
             ),
@@ -934,10 +744,7 @@ class _LoginScreenState extends State<LoginScreen> {
             child: SafeArea(
               child: Padding(
                 padding: const EdgeInsets.only(left: 20.0),
-                child: SizedBox(
-                    width: 100,
-                    height: 50,
-                    child: Image.asset("images/logo_big.png")),
+                child: SizedBox(width: 100, height: 50, child: Image.asset("images/logo_big.png")),
               ),
             ),
           ),
@@ -981,23 +788,19 @@ class _LoginScreenState extends State<LoginScreen> {
   void _forgotDialog() async {
     await showGeneralDialog(
         context: context,
-        pageBuilder: (BuildContext buildContext, Animation<double> animation,
-            Animation<double> secondaryAnimation) {
+        pageBuilder: (BuildContext buildContext, Animation<double> animation, Animation<double> secondaryAnimation) {
           return SafeArea(
             child: Builder(builder: (context) {
-              final TextEditingController _forgotPassControl =
-                  TextEditingController();
+              final TextEditingController forgotPassControl = TextEditingController();
               return Center(
                 child: SizedBox(
                     width: 300,
                     height: MediaQuery.of(context).size.height * 0.22,
-                    child: StatefulBuilder(
-                        builder: (context, StateSetter setState) {
+                    child: StatefulBuilder(builder: (context, StateSetter setState) {
                       return Card(
                         color: const Color(0xFF1B1B1B),
                         child: Padding(
-                          padding:
-                              const EdgeInsets.only(left: 10.0, right: 10.0),
+                          padding: const EdgeInsets.only(left: 10.0, right: 10.0),
                           child: Column(
                             mainAxisAlignment: MainAxisAlignment.start,
                             crossAxisAlignment: CrossAxisAlignment.center,
@@ -1021,44 +824,25 @@ class _LoginScreenState extends State<LoginScreen> {
                                   SizedBox(
                                     width: double.infinity,
                                     child: ClipRRect(
-                                      borderRadius: const BorderRadius.all(
-                                          Radius.circular(5.0)),
+                                      borderRadius: const BorderRadius.all(Radius.circular(5.0)),
                                       child: Container(
                                         color: Colors.black38,
                                         padding: const EdgeInsets.all(5.0),
                                         child: TextField(
-                                            controller: _forgotPassControl,
-                                            style: Theme.of(context)
-                                                .textTheme
-                                                .bodyText1!
-                                                .copyWith(
-                                                    color: Colors.white,
-                                                    fontSize: 18.0),
+                                            controller: forgotPassControl,
+                                            style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.white, fontSize: 18.0),
                                             autocorrect: false,
                                             textAlign: TextAlign.center,
                                             decoration: InputDecoration(
                                               isDense: false,
-                                              contentPadding:
-                                                  const EdgeInsets.only(
-                                                      left: 10.0, bottom: 5.0),
-                                              hintStyle: Theme.of(context)
-                                                  .textTheme
-                                                  .subtitle1!
-                                                  .copyWith(
-                                                      color: Colors.white54,
-                                                      fontSize: 14.0),
-                                              hintText:
-                                                  AppLocalizations.of(context)!
-                                                      .e_mail,
-                                              enabledBorder:
-                                                  const UnderlineInputBorder(
-                                                borderSide: BorderSide(
-                                                    color: Colors.transparent),
+                                              contentPadding: const EdgeInsets.only(left: 10.0, bottom: 5.0),
+                                              hintStyle: Theme.of(context).textTheme.subtitle1!.copyWith(color: Colors.white54, fontSize: 14.0),
+                                              hintText: AppLocalizations.of(context)!.e_mail,
+                                              enabledBorder: const UnderlineInputBorder(
+                                                borderSide: BorderSide(color: Colors.transparent),
                                               ),
-                                              focusedBorder:
-                                                  const UnderlineInputBorder(
-                                                borderSide: BorderSide(
-                                                    color: Colors.transparent),
+                                              focusedBorder: const UnderlineInputBorder(
+                                                borderSide: BorderSide(color: Colors.transparent),
                                               ),
                                             )),
                                       ),
@@ -1067,8 +851,7 @@ class _LoginScreenState extends State<LoginScreen> {
                                 ],
                               ),
                               Padding(
-                                padding: const EdgeInsets.fromLTRB(
-                                    15.0, 22.0, 15.0, 0.0),
+                                padding: const EdgeInsets.fromLTRB(15.0, 22.0, 15.0, 0.0),
                                 child: Row(
                                   mainAxisAlignment: MainAxisAlignment.center,
                                   children: [
@@ -1076,16 +859,13 @@ class _LoginScreenState extends State<LoginScreen> {
                                       width: 90.0,
                                       child: NeuButton(
                                         onTap: () {
-                                          _forgotPass(_forgotPassControl.text);
+                                          _forgotPass(forgotPassControl.text);
                                         },
                                         child: Padding(
                                           padding: const EdgeInsets.all(8.0),
                                           child: Text(
                                             'OK',
-                                            style: Theme.of(context)
-                                                .textTheme
-                                                .headline4!
-                                                .copyWith(color: Colors.white),
+                                            style: Theme.of(context).textTheme.headline4!.copyWith(color: Colors.white),
                                             textAlign: TextAlign.start,
                                           ),
                                         ),
@@ -1104,18 +884,17 @@ class _LoginScreenState extends State<LoginScreen> {
           );
         },
         barrierDismissible: true,
-        barrierLabel:
-            MaterialLocalizations.of(context).modalBarrierDismissLabel,
+        barrierLabel: MaterialLocalizations.of(context).modalBarrierDismissLabel,
         transitionDuration: const Duration(milliseconds: 150));
   }
 
   void _launchURL(String url) async {
-    var _url = url.replaceAll("{0}", "");
-    print(_url);
+    var kUrl = url.replaceAll("{0}", "");
+
     try {
-      await launch(_url);
+      await launchUrl(Uri.parse(kUrl));
     } catch (e) {
-      print(e);
+      if (kDebugMode) print(e);
     }
   }
 }
