@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:decimal/decimal.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'package:auto_size_text/auto_size_text.dart';
@@ -80,7 +81,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
                                 child: SizedBox(
                                   width: 140,
                                   child: AutoSizeText(
-                                    AppLocalizations.of(context)!.sentTo + " " + _formatTx(widget.data.toAddress!),
+                                    "${AppLocalizations.of(context)!.sentTo} ${_formatTx(widget.data.toAddress!)}",
                                     style: Theme.of(context).textTheme.headline3,
                                     maxLines: 1,
                                     minFontSize: 8,
@@ -144,7 +145,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
                                     child: AutoSizeText(
                                       // widget.free!.toString(),
                                       widget.price != null ?
-                                      "-" + (widget.price!.prices!.usd! * Decimal.parse(widget.data.amount!.toString())).toStringAsFixed(3) + " USD": "No Price Data",
+                                      "-${(widget.price!.prices!.usd! * Decimal.parse(widget.data.amount!.toString())).toStringAsFixed(3)} USD": "No Price Data",
                                       style: Theme.of(context).textTheme.headline4!.copyWith(color: const Color(0xffEA3913)),
                                       minFontSize: 8,
                                       maxLines: 1,
@@ -166,7 +167,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
                                     child: AutoSizeText(
                                       // widget.coin.priceData!.prices!.usd!.toStringAsFixed(2) + "\$",
                                       // widget.coin.free!.toStringAsFixed(3),
-                                      widget.data.amount.toString() + " " + widget.data.coin!.name!,
+                                      "${widget.data.amount} ${widget.data.coin!.name!}",
                                       style: Theme.of(context).textTheme.headline3,
                                       maxLines: 1,
                                       minFontSize: 8,
@@ -219,7 +220,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
   String _formatTx(String s) {
     var firstPart = s.substring(0,3);
     var lastPart = s.substring(s.length - 3);
-    return firstPart + "..." + lastPart;
+    return "$firstPart...$lastPart";
   }
   String _getMeDate(String? d) {
     if (d == null) return "";
@@ -229,12 +230,13 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
   }
 
   void _launchURL(String url) async {
-    var _url = url.replaceAll("{0}", "");
-    print(_url);
+    var kUrl = url.replaceAll("{0}", "");
     try {
-      await launch(_url);
+      await launchUrl(Uri.parse(kUrl));
     } catch (e) {
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
   // void _showDetails(TransactionData td) async {
@@ -254,7 +256,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
             Animation<double> secondaryAnimation) {
           return SafeArea(
             child: Builder(builder: (context) {
-              var _copied = false;
+              var copied = false;
               return Center(
                 child: SizedBox(
                     width: 300,
@@ -313,7 +315,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
                                         onTap: () {
                                           Clipboard.setData(ClipboardData(text: td.transactionId.toString()));
                                           setState(() {
-                                            _copied = true;
+                                            copied = true;
                                           });
                                         },
                                         child: ClipRRect(
@@ -331,7 +333,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
                                   IgnorePointer(
                                     child: AnimatedOpacity(
                                       duration:const Duration(milliseconds: 300) ,
-                                      opacity: _copied ? 1.0 : 0.0,
+                                      opacity: copied ? 1.0 : 0.0,
                                       child: SizedBox(
                                         width: double.infinity,
                                         child:
@@ -340,7 +342,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
                                           child: Container(
                                             color: Colors.green,
                                             padding: const EdgeInsets.all(5.0),
-                                            child: Text(AppLocalizations.of(context)!.txcopy+"\n",
+                                            child: Text("${AppLocalizations.of(context)!.txcopy}\n",
                                               textAlign: TextAlign.center,
                                               style: Theme.of(context).textTheme.bodyText1!.copyWith(fontSize: 12.0),
                                             ),
@@ -360,7 +362,7 @@ class _CoinWithdrawalViewState extends State<CoinWithdrawalView> {
                                         onTap: () {
                                           Clipboard.setData(ClipboardData(text: td.transactionId.toString()));
                                           setState(() {
-                                            _copied = true;
+                                            copied = true;
                                           });
                                         },
                                         child: SizedBox(
