@@ -1,5 +1,3 @@
-import 'dart:convert';
-import 'dart:developer';
 import 'dart:io';
 
 import 'package:auto_size_text/auto_size_text.dart';
@@ -7,7 +5,6 @@ import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_phoenix/flutter_phoenix.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:progress_indicators/progress_indicators.dart';
 import 'package:rocketbot/bloc/balance_bloc.dart';
 import 'package:rocketbot/cache/price_graph_cache.dart';
@@ -16,9 +13,6 @@ import 'package:rocketbot/models/balance_list.dart';
 import 'package:rocketbot/models/get_withdraws.dart';
 import 'package:rocketbot/models/pgwid.dart';
 import 'package:rocketbot/models/pos_coins_list.dart';
-import 'package:rocketbot/models/withdraw_confirm.dart';
-import 'package:rocketbot/models/withdraw_pwid.dart';
-import 'package:rocketbot/netInterface/app_exception.dart';
 import 'package:rocketbot/netinterface/api_response.dart';
 import 'package:rocketbot/netinterface/interface.dart';
 import 'package:rocketbot/screens/about_screen.dart';
@@ -36,8 +30,7 @@ import '../widgets/coin_list_view.dart';
 import 'package:rocketbot/support/globals.dart' as globals;
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
-final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-    FlutterLocalNotificationsPlugin();
+final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
 
 class PortfolioScreen extends StatefulWidget {
   const PortfolioScreen({
@@ -68,8 +61,6 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
   bool _paused = false;
   bool _pinEnabled = false;
   bool _hideZero = false;
-
-  double _listHeight = 0.0;
 
   var _dropValue = "Default";
   var _dropValues = ["Default", "By amount", "Alphabetically"];
@@ -165,10 +156,8 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
   }
 
   _posHandle() async {
-   await _lostPosTX();
+    await _lostPosTX();
   }
-
-
 
   _lostPosTX() async {
     List<PGWIdentifier> l = await AppDatabase().getUnfinishedTX();
@@ -181,12 +170,9 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
         await Future.doWhile(() async {
           try {
             await Future.delayed(const Duration(seconds: 3));
-            final _withdrawals = await _interface.get(
-                "Transfers/GetWithdraws?page=1&pageSize=10&coinId=" +
-                    coindID.toString());
-            List<DataWithdrawals>? _with =
-                WithdrawalsModels.fromJson(_withdrawals).data;
-            for (var el in _with!) {
+            final withdrawals = await _interface.get("Transfers/GetWithdraws?page=1&pageSize=10&coinId=$coindID");
+            List<DataWithdrawals>? withRld = WithdrawalsModels.fromJson(withdrawals).data;
+            for (var el in withRld!) {
               if (el.pgwIdentifier! == pgwid) {
                 if (el.transactionId != null) {
                   txid = el.transactionId;
@@ -225,364 +211,305 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
               onRefresh: _refreshData,
               child: SingleChildScrollView(
                 controller: _scrollController,
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.only(
-                            left: 20.0, top: 10.0, bottom: 0.0),
-                        child: Row(
-                          children: [
-                            Text(AppLocalizations.of(context)!.portfolio,
-                                style: Theme.of(context).textTheme.headline4),
-                            const SizedBox(
-                              width: 50,
-                            ),
-                            // SizedBox(
-                            //     height: 30,
-                            //     child: TimeRangeSwitcher(
-                            //       changeTime: _changeTime,
-                            //     )),
-                            Expanded(
-                              child: Align(
-                                alignment: Alignment.centerRight,
-                                child: Padding(
-                                  padding: const EdgeInsets.only(right: 8.0),
-                                  child: SizedBox(
-                                    height: 30,
-                                    width: 25,
-                                    child: NeuButton(
-                                      onTap: () async {
-                                        setState(() {
-                                          if (popMenu) {
-                                            popMenu = false;
-                                          } else {
-                                            popMenu = true;
-                                          }
-                                        });
-                                        // await const FlutterSecureStorage().delete(key: "token");
-                                        // Navigator.of(context).pushReplacement(
-                                        //     PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
-                                        //       return const LoginScreen();
-                                        //     }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
-                                        //       return FadeTransition(opacity: animation, child: child);
-                                        //     }));
-                                      },
-                                      icon: Icon(
-                                        Icons.more_vert,
-                                        color: _socialsOK
-                                            ? Colors.white70
-                                            : Colors.red,
-                                      ),
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  children: [
+                    Padding(
+                      padding: const EdgeInsets.only(left: 20.0, top: 10.0, bottom: 0.0),
+                      child: Row(
+                        children: [
+                          Text(AppLocalizations.of(context)!.portfolio, style: Theme.of(context).textTheme.headline4),
+                          const SizedBox(
+                            width: 50,
+                          ),
+                          // SizedBox(
+                          //     height: 30,
+                          //     child: TimeRangeSwitcher(
+                          //       changeTime: _changeTime,
+                          //     )),
+                          Expanded(
+                            child: Align(
+                              alignment: Alignment.centerRight,
+                              child: Padding(
+                                padding: const EdgeInsets.only(right: 8.0),
+                                child: SizedBox(
+                                  height: 30,
+                                  width: 25,
+                                  child: NeuButton(
+                                    onTap: () async {
+                                      setState(() {
+                                        if (popMenu) {
+                                          popMenu = false;
+                                        } else {
+                                          popMenu = true;
+                                        }
+                                      });
+                                      // await const FlutterSecureStorage().delete(key: "token");
+                                      // Navigator.of(context).pushReplacement(
+                                      //     PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
+                                      //       return const LoginScreen();
+                                      //     }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
+                                      //       return FadeTransition(opacity: animation, child: child);
+                                      //     }));
+                                    },
+                                    icon: Icon(
+                                      Icons.more_vert,
+                                      color: _socialsOK ? Colors.white70 : Colors.red,
                                     ),
                                   ),
                                 ),
                               ),
-                            )
-                          ],
-                        ),
+                            ),
+                          )
+                        ],
                       ),
-                      SizedBox(
-                        width: double.infinity,
-                        height: 250,
-                        child: portCalc
-                            ? Stack(
-                                // alignment: AlignmentDirectional.center,
-                                children: [
-                                  Padding(
-                                    padding: const EdgeInsets.only(
-                                        left: 0.0, right: 10.0, top: 50.0),
-                                    child: SizedBox(
-                                      width: double.infinity,
-                                      child: Transform(
-                                        transform: scaleXYZTransform(),
-                                        child: const Image(
-                                          fit: BoxFit.fitWidth,
-                                          image: AssetImage("images/wave.png"),
-                                        ),
-                                      ),
-                                    ),
-                                  ),
-                                  Padding(
-                                    padding: const EdgeInsets.only(
-                                        left: 170.0, right: 0.0, top: 90.0),
-                                    child: Transform.scale(
-                                      scale: 0.35,
+                    ),
+                    SizedBox(
+                      width: double.infinity,
+                      height: 250,
+                      child: portCalc
+                          ? Stack(
+                              // alignment: AlignmentDirectional.center,
+                              children: [
+                                Padding(
+                                  padding: const EdgeInsets.only(left: 0.0, right: 10.0, top: 50.0),
+                                  child: SizedBox(
+                                    width: double.infinity,
+                                    child: Transform(
+                                      transform: scaleXYZTransform(),
                                       child: const Image(
-                                        image:
-                                            AssetImage("images/rocket_pin.png"),
+                                        fit: BoxFit.fitWidth,
+                                        image: AssetImage("images/wave.png"),
                                       ),
                                     ),
                                   ),
-                                  const Align(
-                                    alignment: Alignment.center,
-                                    child: Padding(
-                                      padding: EdgeInsets.only(bottom: 98.0),
-                                      child: AspectRatio(
-                                        aspectRatio: 1.6,
-                                        child: Image(
-                                            fit: BoxFit.fitWidth,
-                                            image: AssetImage(
-                                                "images/price_frame.png")),
-                                      ),
+                                ),
+                                Padding(
+                                  padding: const EdgeInsets.only(left: 170.0, right: 0.0, top: 90.0),
+                                  child: Transform.scale(
+                                    scale: 0.35,
+                                    child: const Image(
+                                      image: AssetImage("images/rocket_pin.png"),
                                     ),
                                   ),
-                                  Padding(
-                                    padding: const EdgeInsets.only(
-                                        bottom: 130.0, top: 25.0),
-                                    child: Center(
-                                      child: Column(
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.center,
-                                        children: [
-                                          SizedBox(
-                                            width: 200,
-                                            child: AutoSizeText(
-                                              "\$" +
-                                                  totalUSD.toStringAsFixed(2),
-                                              style: Theme.of(context)
-                                                  .textTheme
-                                                  .headline1,
-                                              minFontSize: 8.0,
-                                              maxLines: 1,
-                                              textAlign: TextAlign.center,
-                                            ),
+                                ),
+                                const Align(
+                                  alignment: Alignment.center,
+                                  child: Padding(
+                                    padding: EdgeInsets.only(bottom: 98.0),
+                                    child: AspectRatio(
+                                      aspectRatio: 1.6,
+                                      child: Image(fit: BoxFit.fitWidth, image: AssetImage("images/price_frame.png")),
+                                    ),
+                                  ),
+                                ),
+                                Padding(
+                                  padding: const EdgeInsets.only(bottom: 130.0, top: 25.0),
+                                  child: Center(
+                                    child: Column(
+                                      mainAxisAlignment: MainAxisAlignment.center,
+                                      crossAxisAlignment: CrossAxisAlignment.center,
+                                      children: [
+                                        SizedBox(
+                                          width: 200,
+                                          child: AutoSizeText(
+                                            "\$${totalUSD.toStringAsFixed(2)}",
+                                            style: Theme.of(context).textTheme.headline1,
+                                            minFontSize: 8.0,
+                                            maxLines: 1,
+                                            textAlign: TextAlign.center,
                                           ),
-                                          const SizedBox(
-                                            height: 3.0,
-                                          ),
-                                          SizedBox(
-                                            width: 130,
-                                            child: AutoSizeText(
-                                              _formatPrice(totalBTC) + " BTC",
-                                              style: Theme.of(context)
-                                                  .textTheme
-                                                  .headline2,
-                                              minFontSize: 8.0,
-                                              maxLines: 1,
-                                              textAlign: TextAlign.center,
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                  ),
-                                ],
-                              )
-                            : Container(),
-                      ),
-                      const SizedBox(
-                        height: 5.0,
-                      ),
-                      SizedBox(
-                        height: 25.0,
-                        width: double.infinity,
-                        child: Align(
-                          alignment: Alignment.center,
-                          child: Padding(
-                            padding:
-                                const EdgeInsets.only(left: 8.0, right: 3.0),
-                            child: Opacity(
-                              opacity: 0.6,
-                              child: Row(
-                                mainAxisAlignment: MainAxisAlignment.start,
-                                crossAxisAlignment: CrossAxisAlignment.center,
-                                children: [
-                                  const Padding(
-                                    padding: EdgeInsets.only(bottom: 0.0),
-                                    child: Icon(
-                                      Icons.sort,
-                                      color: Colors.white30,
-                                      size: 10.0,
-                                    ),
-                                  ),
-                                  // Text('sort by:', style: Theme.of(context).textTheme.headline2!.copyWith( fontSize: 14.0, color: Colors.white30)),
-                                  const SizedBox(
-                                    width: 5.0,
-                                  ),
-                                  DropdownButtonHideUnderline(
-                                    child: DropdownButton<String>(
-                                      value: _dropValue,
-                                      isDense: true,
-                                      onChanged: (String? val) async {
-                                        _bloc!.showWait();
-                                        setState(() {
-                                          _dropValue = val!;
-                                          _listCoins = null;
-                                        });
-                                        await Future.delayed(const Duration(milliseconds: 100), () {});
-                                        int sort = _dropValues.indexWhere(
-                                            (element) => element == _dropValue);
-                                        await SecureStorage.writeStorage(
-                                            key: globals.SORT_TYPE,
-                                            value: sort.toString());
-                                        await _bloc!
-                                            .fetchBalancesList(sort: sort);
-                                        await _checkZero();
-                                      },
-                                      items: _dropValues
-                                          .map((e) => DropdownMenuItem(
-                                              value: e,
-                                              child: SizedBox(
-                                                  width: 130,
-                                                  child: Padding(
-                                                    padding:
-                                                        const EdgeInsets.only(
-                                                            bottom: 2.0),
-                                                    child: Text(e,
-                                                        style: Theme.of(context)
-                                                            .textTheme
-                                                            .headline2!
-                                                            .copyWith(
-                                                                fontSize: 11.0,
-                                                                color: Colors
-                                                                    .white70)),
-                                                  ))))
-                                          .toList(),
-                                    ),
-                                  ),
-                                  Expanded(
-                                    child: Padding(
-                                      padding: const EdgeInsets.only(
-                                          left: 5.0, right: 8.0, top: 1.0),
-                                      child: SizedBox(
-                                        height: 0.5,
-                                        child: Container(
-                                          color: portCalc
-                                              ? Colors.white12
-                                              : Colors.transparent,
                                         ),
+                                        const SizedBox(
+                                          height: 3.0,
+                                        ),
+                                        SizedBox(
+                                          width: 130,
+                                          child: AutoSizeText(
+                                            "${_formatPrice(totalBTC)} BTC",
+                                            style: Theme.of(context).textTheme.headline2,
+                                            minFontSize: 8.0,
+                                            maxLines: 1,
+                                            textAlign: TextAlign.center,
+                                          ),
+                                        ),
+                                      ],
+                                    ),
+                                  ),
+                                ),
+                              ],
+                            )
+                          : Container(),
+                    ),
+                    const SizedBox(
+                      height: 5.0,
+                    ),
+                    SizedBox(
+                      height: 25.0,
+                      width: double.infinity,
+                      child: Align(
+                        alignment: Alignment.center,
+                        child: Padding(
+                          padding: const EdgeInsets.only(left: 8.0, right: 3.0),
+                          child: Opacity(
+                            opacity: 0.6,
+                            child: Row(
+                              mainAxisAlignment: MainAxisAlignment.start,
+                              crossAxisAlignment: CrossAxisAlignment.center,
+                              children: [
+                                const Padding(
+                                  padding: EdgeInsets.only(bottom: 0.0),
+                                  child: Icon(
+                                    Icons.sort,
+                                    color: Colors.white30,
+                                    size: 10.0,
+                                  ),
+                                ),
+                                // Text('sort by:', style: Theme.of(context).textTheme.headline2!.copyWith( fontSize: 14.0, color: Colors.white30)),
+                                const SizedBox(
+                                  width: 5.0,
+                                ),
+                                DropdownButtonHideUnderline(
+                                  child: DropdownButton<String>(
+                                    value: _dropValue,
+                                    isDense: true,
+                                    onChanged: (String? val) async {
+                                      _bloc!.showWait();
+                                      setState(() {
+                                        _dropValue = val!;
+                                        _listCoins = null;
+                                      });
+                                      await Future.delayed(const Duration(milliseconds: 100), () {});
+                                      int sort = _dropValues.indexWhere((element) => element == _dropValue);
+                                      await SecureStorage.writeStorage(key: globals.SORT_TYPE, value: sort.toString());
+                                      await _bloc!.fetchBalancesList(sort: sort);
+                                      await _checkZero();
+                                    },
+                                    items: _dropValues
+                                        .map((e) => DropdownMenuItem(
+                                            value: e,
+                                            child: SizedBox(
+                                                width: 130,
+                                                child: Padding(
+                                                  padding: const EdgeInsets.only(bottom: 2.0),
+                                                  child: Text(e,
+                                                      style: Theme.of(context).textTheme.headline2!.copyWith(fontSize: 11.0, color: Colors.white70)),
+                                                ))))
+                                        .toList(),
+                                  ),
+                                ),
+                                Expanded(
+                                  child: Padding(
+                                    padding: const EdgeInsets.only(left: 5.0, right: 8.0, top: 1.0),
+                                    child: SizedBox(
+                                      height: 0.5,
+                                      child: Container(
+                                        color: portCalc ? Colors.white12 : Colors.transparent,
                                       ),
                                     ),
                                   ),
-                                  const SizedBox(
-                                    width: 5.0,
-                                  ),
-                                  Padding(
-                                    padding: const EdgeInsets.only(bottom: 1.0),
-                                    child: FlatCustomButton(
-                                      onTap: () async {
-                                        _bloc!.showWait();
-                                        await Future.delayed(const Duration(milliseconds: 100), () {});
-                                        setState(() {
-                                          _listCoins = null;
-                                          if (_hideZero) {
-                                            _hideZero = false;
-                                            _bloc!.fetchBalancesList();
-                                            // _bloc!.filterCoinsList(zero: _hideZero);
-                                          } else {
-                                            _hideZero = true;
-                                            _bloc!.filterCoinsList(
-                                                zero: _hideZero);
-                                            // _bloc!.filterCoinsList(zero: _hideZero);
-                                          }
-                                        });
-                                      },
-                                      child: Padding(
-                                        padding: const EdgeInsets.only(
-                                            left: 0.0, right: 1.0),
-                                        child: FittedBox(
-                                            child: AutoSizeText(
-                                          AppLocalizations.of(context)!
-                                              .hide_zeros,
-                                          style: Theme.of(context)
-                                              .textTheme
-                                              .headline2!
-                                              .copyWith(
-                                                  fontSize: 11.0,
-                                                  color: _hideZero
-                                                      ? Colors.white
-                                                      : Colors.white30),
-                                        )),
-                                      ),
+                                ),
+                                const SizedBox(
+                                  width: 5.0,
+                                ),
+                                Padding(
+                                  padding: const EdgeInsets.only(bottom: 1.0),
+                                  child: FlatCustomButton(
+                                    onTap: () async {
+                                      _bloc!.showWait();
+                                      await Future.delayed(const Duration(milliseconds: 100), () {});
+                                      setState(() {
+                                        _listCoins = null;
+                                        if (_hideZero) {
+                                          _hideZero = false;
+                                          _bloc!.fetchBalancesList();
+                                          // _bloc!.filterCoinsList(zero: _hideZero);
+                                        } else {
+                                          _hideZero = true;
+                                          _bloc!.filterCoinsList(zero: _hideZero);
+                                          // _bloc!.filterCoinsList(zero: _hideZero);
+                                        }
+                                      });
+                                    },
+                                    child: Padding(
+                                      padding: const EdgeInsets.only(left: 0.0, right: 1.0),
+                                      child: FittedBox(
+                                          child: AutoSizeText(
+                                        AppLocalizations.of(context)!.hide_zeros,
+                                        style: Theme.of(context)
+                                            .textTheme
+                                            .headline2!
+                                            .copyWith(fontSize: 11.0, color: _hideZero ? Colors.white : Colors.white30),
+                                      )),
                                     ),
                                   ),
-                                  const SizedBox(
-                                    width: 8.0,
-                                  ),
-                                ],
-                              ),
+                                ),
+                                const SizedBox(
+                                  width: 8.0,
+                                ),
+                              ],
                             ),
                           ),
                         ),
                       ),
-                      Flexible(
-                        child: Padding(
-                          padding: const EdgeInsets.only(top: 3.0),
-                          child: StreamBuilder<ApiResponse<List<CoinBalance>>>(
-                            stream: _bloc!.coinsListStream,
-                            builder: (context, snapshot) {
-                              if (snapshot.hasData) {
-                                switch (snapshot.data!.status) {
-                                  case Status.LOADING:
-                                    _listCoins = null;
-                                    return Align(
-                                      alignment: Alignment.topCenter,
-                                      child: Padding(
-                                        padding:
-                                            const EdgeInsets.only(top: 40.0),
-                                        child: HeartbeatProgressIndicator(
-                                          startScale: 0.01,
-                                          endScale: 0.2,
-                                          child: const Image(
-                                            image: AssetImage(
-                                                'images/rocketbot_logo.png'),
-                                            color: Colors.white30,
-                                          ),
+                    ),
+                    Flexible(
+                      child: Padding(
+                        padding: const EdgeInsets.only(top: 3.0),
+                        child: StreamBuilder<ApiResponse<List<CoinBalance>>>(
+                          stream: _bloc!.coinsListStream,
+                          builder: (context, snapshot) {
+                            if (snapshot.hasData) {
+                              switch (snapshot.data!.status) {
+                                case Status.loading:
+                                  _listCoins = null;
+                                  return Align(
+                                    alignment: Alignment.topCenter,
+                                    child: Padding(
+                                      padding: const EdgeInsets.only(top: 40.0),
+                                      child: HeartbeatProgressIndicator(
+                                        startScale: 0.01,
+                                        endScale: 0.2,
+                                        child: const Image(
+                                          image: AssetImage('images/rocketbot_logo.png'),
+                                          color: Colors.white30,
                                         ),
                                       ),
-                                    );
-                                  case Status.COMPLETED:
-                                    PriceGraphCache.refreshAllRecords();
-                                    if (_listCoins == null) {
-                                      _listCoins = snapshot.data!.data!;
-                                      _listHeight = _hideZero
-                                          ? _listCoins!.length *
-                                              MediaQuery.of(context)
-                                                  .size
-                                                  .height *
-                                              0.11
-                                          : _listCoins!.length *
-                                              (MediaQuery.of(context)
-                                                      .size
-                                                      .height *
-                                                  0.085);
-                                      // widget.passBalances(listCoins);
-                                      _calculatePortfolio();
-                                    }
-                                    return ListView.builder(
-                                        shrinkWrap: true,
-                                        physics:
-                                            const NeverScrollableScrollPhysics(),
-                                        itemCount: snapshot.data!.data!.length,
-                                        itemBuilder: (ctx, index) {
-                                          return CoinListView(
-                                            key: ValueKey(snapshot
-                                                .data!.data![index].coin!.id!),
-                                            coin: snapshot.data!.data![index],
-                                            free: Decimal.parse(snapshot
-                                                .data!.data![index].free
-                                                .toString()),
-                                            coinSwitch: _changeCoin,
-                                          );
-                                        });
-                                  case Status.ERROR:
-                                    debugPrint("error");
-                                    break;
-                                }
+                                    ),
+                                  );
+                                case Status.completed:
+                                  PriceGraphCache.refreshAllRecords();
+                                  if (_listCoins == null) {
+                                    _listCoins = snapshot.data!.data!;
+                                    _calculatePortfolio();
+                                  }
+                                  return ListView.builder(
+                                      shrinkWrap: true,
+                                      physics: const NeverScrollableScrollPhysics(),
+                                      itemCount: snapshot.data!.data!.length,
+                                      itemBuilder: (ctx, index) {
+                                        return CoinListView(
+                                          key: ValueKey(snapshot.data!.data![index].coin!.id!),
+                                          coin: snapshot.data!.data![index],
+                                          free: Decimal.parse(snapshot.data!.data![index].free.toString()),
+                                          coinSwitch: _changeCoin,
+                                        );
+                                      });
+                                case Status.error:
+                                  debugPrint("error");
+                                  break;
                               }
-                              return Container();
-                            },
-                          ),
+                            }
+                            return Container();
+                          },
                         ),
                       ),
-                    ],
-                  ),
+                    ),
+                  ],
                 ),
               ),
+            ),
             // ),
             Visibility(
                 visible: false,
@@ -609,8 +536,7 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
                     elevation: 10.0,
                     color: Colors.transparent,
                     child: ClipRRect(
-                      borderRadius:
-                          const BorderRadius.all(Radius.circular(8.0)),
+                      borderRadius: const BorderRadius.all(Radius.circular(8.0)),
                       child: Container(
                         width: 120,
                         color: const Color(0xFF1B1B1B),
@@ -625,64 +551,38 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
                                     child: SizedBox(
                                       width: 140,
                                       child: TextButton(
-                                        child: Text(
-                                          AppLocalizations.of(context)!
-                                                  .socials_popup +
-                                              (_socialsOK ? '' : ' (!)'),
-                                          style: Theme.of(context)
-                                              .textTheme
-                                              .headline1!
-                                              .copyWith(
-                                                  fontSize: 14.0,
-                                                  color: _socialsOK
-                                                      ? Colors.white
-                                                      : Colors.red),
-                                        ),
                                         style: ButtonStyle(
-                                            backgroundColor:
-                                                MaterialStateProperty.resolveWith(
-                                                    (states) =>
-                                                        qrColors(states)),
-                                            shape: MaterialStateProperty.all<
-                                                    RoundedRectangleBorder>(
-                                                RoundedRectangleBorder(
-                                                    borderRadius:
-                                                        BorderRadius.circular(
-                                                            0.0),
-                                                    side: const BorderSide(
-                                                        color: Colors
-                                                            .transparent)))),
+                                            backgroundColor: MaterialStateProperty.resolveWith((states) => qrColors(states)),
+                                            shape: MaterialStateProperty.all<RoundedRectangleBorder>(RoundedRectangleBorder(
+                                                borderRadius: BorderRadius.circular(0.0), side: const BorderSide(color: Colors.transparent)))),
                                         onPressed: () {
                                           setState(() {
                                             popMenu = false;
                                           });
                                           Navigator.of(context)
-                                              .push(PageRouteBuilder(
-                                                  pageBuilder:
-                                                      (BuildContext context, _,
-                                                          __) {
+                                              .push(PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
                                                 return SocialScreen(
                                                   socials: _socials,
                                                   me: _me!,
                                                 );
-                                              }, transitionsBuilder: (_,
-                                                      Animation<double>
-                                                          animation,
-                                                      __,
-                                                      Widget child) {
-                                                return FadeTransition(
-                                                    opacity: animation,
-                                                    child: child);
+                                              }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
+                                                return FadeTransition(opacity: animation, child: child);
                                               }))
                                               .then((value) => _getUserInfo());
                                         },
+                                        child: Text(
+                                          AppLocalizations.of(context)!.socials_popup + (_socialsOK ? '' : ' (!)'),
+                                          style: Theme.of(context)
+                                              .textTheme
+                                              .headline1!
+                                              .copyWith(fontSize: 14.0, color: _socialsOK ? Colors.white : Colors.red),
+                                        ),
                                       ),
                                     ),
                                   ),
                                 )),
                             Padding(
-                              padding:
-                                  const EdgeInsets.only(left: 4.0, right: 4.0),
+                              padding: const EdgeInsets.only(left: 4.0, right: 4.0),
                               child: Container(
                                 height: 0.5,
                                 color: Colors.white12,
@@ -697,56 +597,32 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
                                     child: SizedBox(
                                       width: 140,
                                       child: TextButton(
-                                        child: Text(
-                                          AppLocalizations.of(context)!
-                                              .settings_popup,
-                                          style: Theme.of(context)
-                                              .textTheme
-                                              .headline1!
-                                              .copyWith(fontSize: 14.0),
-                                        ),
                                         style: ButtonStyle(
-                                            backgroundColor:
-                                                MaterialStateProperty.resolveWith(
-                                                    (states) =>
-                                                        qrColors(states)),
-                                            shape: MaterialStateProperty.all<
-                                                    RoundedRectangleBorder>(
-                                                RoundedRectangleBorder(
-                                                    borderRadius:
-                                                        BorderRadius.circular(
-                                                            0.0),
-                                                    side: const BorderSide(
-                                                        color: Colors
-                                                            .transparent)))),
+                                            backgroundColor: MaterialStateProperty.resolveWith((states) => qrColors(states)),
+                                            shape: MaterialStateProperty.all<RoundedRectangleBorder>(RoundedRectangleBorder(
+                                                borderRadius: BorderRadius.circular(0.0), side: const BorderSide(color: Colors.transparent)))),
                                         onPressed: () {
                                           setState(() {
                                             popMenu = false;
                                           });
                                           Navigator.of(context)
-                                              .push(PageRouteBuilder(
-                                                  pageBuilder:
-                                                      (BuildContext context, _,
-                                                          __) {
+                                              .push(PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
                                                 return const SettingsScreen();
-                                              }, transitionsBuilder: (_,
-                                                      Animation<double>
-                                                          animation,
-                                                      __,
-                                                      Widget child) {
-                                                return FadeTransition(
-                                                    opacity: animation,
-                                                    child: child);
+                                              }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
+                                                return FadeTransition(opacity: animation, child: child);
                                               }))
                                               .then((value) => _fillSort());
                                         },
+                                        child: Text(
+                                          AppLocalizations.of(context)!.settings_popup,
+                                          style: Theme.of(context).textTheme.headline1!.copyWith(fontSize: 14.0),
+                                        ),
                                       ),
                                     ),
                                   ),
                                 )),
                             Padding(
-                              padding:
-                                  const EdgeInsets.only(left: 4.0, right: 4.0),
+                              padding: const EdgeInsets.only(left: 4.0, right: 4.0),
                               child: Container(
                                 height: 0.5,
                                 color: Colors.white12,
@@ -761,46 +637,24 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
                                     child: SizedBox(
                                       width: 140,
                                       child: TextButton(
-                                        child: Text(
-                                          AppLocalizations.of(context)!
-                                              .about_popup,
-                                          style: Theme.of(context)
-                                              .textTheme
-                                              .headline1!
-                                              .copyWith(fontSize: 14.0),
-                                        ),
                                         style: ButtonStyle(
-                                            backgroundColor:
-                                                MaterialStateProperty.resolveWith(
-                                                    (states) =>
-                                                        qrColors(states)),
-                                            shape: MaterialStateProperty.all<
-                                                    RoundedRectangleBorder>(
-                                                RoundedRectangleBorder(
-                                                    borderRadius:
-                                                        BorderRadius.circular(
-                                                            0.0),
-                                                    side: const BorderSide(
-                                                        color: Colors
-                                                            .transparent)))),
+                                            backgroundColor: MaterialStateProperty.resolveWith((states) => qrColors(states)),
+                                            shape: MaterialStateProperty.all<RoundedRectangleBorder>(RoundedRectangleBorder(
+                                                borderRadius: BorderRadius.circular(0.0), side: const BorderSide(color: Colors.transparent)))),
                                         onPressed: () {
                                           setState(() {
                                             popMenu = false;
                                           });
-                                          Navigator.of(context).push(
-                                              PageRouteBuilder(pageBuilder:
-                                                  (BuildContext context, _,
-                                                      __) {
+                                          Navigator.of(context).push(PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
                                             return const AboutScreen();
-                                          }, transitionsBuilder: (_,
-                                                  Animation<double> animation,
-                                                  __,
-                                                  Widget child) {
-                                            return FadeTransition(
-                                                opacity: animation,
-                                                child: child);
+                                          }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
+                                            return FadeTransition(opacity: animation, child: child);
                                           }));
                                         },
+                                        child: Text(
+                                          AppLocalizations.of(context)!.about_popup,
+                                          style: Theme.of(context).textTheme.headline1!.copyWith(fontSize: 14.0),
+                                        ),
                                       ),
                                     ),
                                   ),
@@ -842,17 +696,20 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
 
   _changeCoin(CoinBalance c) async {
     pl ??= await _getPosCoins();
-    Navigator.of(context)
-        .push(PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
-      return MainScreen(
-        posCoinsList: pl,
-        refreshList: _refreshData,
-        coinBalance: c,
-        listCoins: _listCoins,
-      );
-    }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
-      return FadeTransition(opacity: animation, child: child);
-    })).then((value) => _lostPosTX());
+    if (mounted) {
+      Navigator.of(context)
+          .push(PageRouteBuilder(pageBuilder: (BuildContext context, _, __) {
+            return MainScreen(
+              posCoinsList: pl,
+              refreshList: _refreshData,
+              coinBalance: c,
+              listCoins: _listCoins,
+            );
+          }, transitionsBuilder: (_, Animation<double> animation, __, Widget child) {
+            return FadeTransition(opacity: animation, child: child);
+          }))
+          .then((value) => _lostPosTX());
+    }
   }
 
   Future<PosCoinsList?> _getPosCoins() async {
@@ -865,11 +722,11 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
   }
 
   String _formatPrice(double d) {
-    var _split = d.toString().split('.');
-    var _decimal = _split[1];
-    if (_decimal.length >= 8) {
-      var _sub = _decimal.substring(0, 7);
-      return _split[0] + "." + _sub;
+    var split = d.toString().split('.');
+    var decimal = split[1];
+    if (decimal.length >= 8) {
+      var sub = decimal.substring(0, 7);
+      return "${split[0]}.$sub";
     } else {
       return d.toString();
     }
@@ -880,14 +737,14 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
       totalUSD = 0;
       totalBTC = 0;
       for (var element in _listCoins!) {
-        double? _freeCoins = element.free;
-        double? _priceUSD = element.priceData?.prices?.usd!.toDouble();
-        double? _priceBTC = element.priceData?.prices?.btc!.toDouble();
-        if (_priceUSD != null && _priceBTC != null) {
-          double _usd = _freeCoins! * _priceUSD;
-          totalUSD += _usd;
-          double _btc = _freeCoins * _priceBTC;
-          totalBTC += _btc;
+        double? freeCoins = element.free;
+        double? priceUSD = element.priceData?.prices?.usd!.toDouble();
+        double? priceBTC = element.priceData?.prices?.btc!.toDouble();
+        if (priceUSD != null && priceBTC != null) {
+          double usd = freeCoins! * priceUSD;
+          totalUSD += usd;
+          double btc = freeCoins * priceBTC;
+          totalBTC += btc;
         }
       }
 
@@ -923,8 +780,7 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
     return s;
   }
 
-  void _onDidReceiveLocalNotification(
-      int id, String? title, String? body, String? payload) {
+  void _onDidReceiveLocalNotification(int id, String? title, String? body, String? payload) {
     Dialogs.openAlertBox(context, "Alert", payload!);
   }
 
@@ -952,8 +808,8 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
   @override
   void onPaused() async {
     if (!_paused) {
-      var _endTime = DateTime.now().millisecondsSinceEpoch + (1000 * 60);
-      await SecureStorage.writeStorage(key: globals.COUNTDOWN, value: _endTime.toString());
+      var endTime = DateTime.now().millisecondsSinceEpoch + (1000 * 60);
+      await SecureStorage.writeStorage(key: globals.COUNTDOWN, value: endTime.toString());
       _paused = true;
     }
   }
@@ -971,10 +827,10 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
   }
 
   Future<bool> _checkCountdown() async {
-    var _countDown = await SecureStorage.readStorage(key: globals.COUNTDOWN);
-    if (_countDown != null) {
+    var countDown = await SecureStorage.readStorage(key: globals.COUNTDOWN);
+    if (countDown != null) {
       int nowDate = DateTime.now().millisecondsSinceEpoch;
-      int countTime = int.parse(_countDown);
+      int countTime = int.parse(countDown);
       if (nowDate > countTime) {
         return true;
       } else {
@@ -998,8 +854,7 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
       );
 
       await flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()!
+          .resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>()!
           .createNotificationChannel(channel);
 
       AndroidNotificationChannel channel2 = const AndroidNotificationChannel(
@@ -1014,21 +869,15 @@ class PortfolioScreenState extends LifecycleWatcherState<PortfolioScreen> {
       );
 
       await flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()!
+          .resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>()!
           .createNotificationChannel(channel2);
     }
-    const AndroidInitializationSettings initializationSettingsAndroid =
-        AndroidInitializationSettings('@mipmap/ic_notification');
+    const AndroidInitializationSettings initializationSettingsAndroid = AndroidInitializationSettings('@mipmap/ic_notification');
     final IOSInitializationSettings initializationSettingsIOS =
-        IOSInitializationSettings(
-            onDidReceiveLocalNotification: _onDidReceiveLocalNotification);
+        IOSInitializationSettings(onDidReceiveLocalNotification: _onDidReceiveLocalNotification);
     final InitializationSettings initializationSettings =
-        InitializationSettings(
-            android: initializationSettingsAndroid,
-            iOS: initializationSettingsIOS);
-    flutterLocalNotificationsPlugin.initialize(initializationSettings,
-        onSelectNotification: _onSelectNotification);
+        InitializationSettings(android: initializationSettingsAndroid, iOS: initializationSettingsIOS);
+    flutterLocalNotificationsPlugin.initialize(initializationSettings, onSelectNotification: _onSelectNotification);
   }
 
   _checkZero() async {
