@@ -2,16 +2,16 @@ extension DurationExtensions on Duration {
   /// Converts the duration into a readable string
   /// 05:15
   String toHoursMinutes() {
-    String twoDigitMinutes = _toTwoDigits(this.inMinutes.remainder(60));
-    return "${_toTwoDigits(this.inHours)}:$twoDigitMinutes";
+    String twoDigitMinutes = _toTwoDigits(inMinutes.remainder(60));
+    return "${_toTwoDigits(inHours)}:$twoDigitMinutes";
   }
 
   /// Converts the duration into a readable string
   /// 05:15:35
   String toHoursMinutesSeconds() {
-    String twoDigitMinutes = _toTwoDigits(this.inMinutes.remainder(60));
-    String twoDigitSeconds = _toTwoDigits(this.inSeconds.remainder(60));
-    return "${_toTwoDigits(this.inHours)}:$twoDigitMinutes:$twoDigitSeconds";
+    String twoDigitMinutes = _toTwoDigits(inMinutes.remainder(60));
+    String twoDigitSeconds = _toTwoDigits(inSeconds.remainder(60));
+    return "${_toTwoDigits(inHours)}:$twoDigitMinutes:$twoDigitSeconds";
   }
 
   String _toTwoDigits(int n) {
