@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 class SecureStorage {
@@ -5,24 +6,28 @@ class SecureStorage {
 
   static Future<String?> readStorage({required String key}) {
     try {
-      const FlutterSecureStorage _storage = FlutterSecureStorage();
+      const FlutterSecureStorage storage = FlutterSecureStorage();
       const optionsApple = IOSOptions(accessibility: IOSAccessibility.first_unlock);
       const optionsAndroid = AndroidOptions(encryptedSharedPreferences: true);
-      return  _storage.read(key: key, iOptions: optionsApple, aOptions: optionsAndroid);
+      return  storage.read(key: key, iOptions: optionsApple, aOptions: optionsAndroid);
     } catch (e) {
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
       return Future.value(null);
     }
   }
 
   static Future<void> writeStorage({required String key,required String value}) {
     try {
-      const FlutterSecureStorage _storage = FlutterSecureStorage();
+      const FlutterSecureStorage storage = FlutterSecureStorage();
       const optionsApple = IOSOptions(accessibility: IOSAccessibility.first_unlock);
       const optionsAndroid = AndroidOptions(encryptedSharedPreferences: true);
-      return  _storage.write(key: key, value: value, iOptions: optionsApple, aOptions: optionsAndroid);
+      return  storage.write(key: key, value: value, iOptions: optionsApple, aOptions: optionsAndroid);
     } catch (e) {
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
       return Future.value(null);
     }
 
@@ -30,12 +35,14 @@ class SecureStorage {
 
   static Future<void> deleteStorage({required String key}) {
     try {
-      const FlutterSecureStorage _storage = FlutterSecureStorage();
+      const FlutterSecureStorage storage = FlutterSecureStorage();
       const optionsApple = IOSOptions(accessibility: IOSAccessibility.first_unlock);
       const optionsAndroid = AndroidOptions(encryptedSharedPreferences: true);
-      return   _storage.delete(key: key, iOptions: optionsApple, aOptions: optionsAndroid);
+      return   storage.delete(key: key, iOptions: optionsApple, aOptions: optionsAndroid);
     } catch (e) {
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
       return Future.value(null);
     }
   }
