@@ -4,7 +4,6 @@ import 'dart:io';
 import 'package:path_provider/path_provider.dart';
 import 'package:http/http.dart' show get;
 
-
 class PictureCacheWidget extends StatefulWidget {
   final Coin? coin;
 
@@ -28,16 +27,14 @@ class _PictureCacheWidgetState extends State<PictureCacheWidget> {
 
   _handleFile() async {
     var documentDirectory = await getApplicationDocumentsDirectory();
-    var firstPath = documentDirectory.path + "/images";
-    var filePathAndName = documentDirectory.path + '/images/${widget.coin!.cryptoId!}';
+    var firstPath = "${documentDirectory.path}/images";
+    var filePathAndName = '${documentDirectory.path}/images/${widget.coin!.cryptoId!}';
     if (await File(filePathAndName).exists()) {
       setState(() {
         _imageFile = File.fromUri(Uri.parse(filePathAndName));
       });
     } else {
-      var response = await get(Uri.parse('https://app.rocketbot.pro/Image?imageId=' +
-          widget.coin!
-              .imageSmallid!));
+      var response = await get(Uri.parse('https://app.rocketbot.pro/Image?imageId=${widget.coin!.imageSmallid!}'));
       await Directory(firstPath).create(recursive: true);
       File file2 = File(filePathAndName);
       file2.writeAsBytesSync(response.bodyBytes);
@@ -49,6 +46,6 @@ class _PictureCacheWidgetState extends State<PictureCacheWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return _imageFile != null ? Image.file(_imageFile!): Container();
+    return _imageFile != null ? Image.file(_imageFile!) : Container();
   }
 }
